requirements.txt
# Python dependencies for Vercel serverless functions
fastapi==0.109.0
python-multipart==0.0.6
PyMuPDF==1.23.22
Pillow==10.2.0
supabase==2.3.0
pydantic==2.5.0
